+++ rust/src/interface/storage.rs
@@ -4,12 +4,12 @@ use alloc::{vec, vec::Vec};
 pub struct StorageInterface {
     pub connected: extern "C" fn() -> bool,
     pub busy: extern "C" fn() -> bool,
-    pub write: extern "C" fn(address: u16, count: u8, buffer: *const u8) -> bool,
-    pub read: extern "C" fn(address: u16, count: u8, buffer: *mut u8) -> bool,
+    pub write: extern "C" fn(address: u16, count: u16, buffer: *const u8) -> bool,
+    pub read: extern "C" fn(address: u16, count: u16, buffer: *mut u8) -> bool,
 }
 
 impl StorageInterface {
-    pub fn read(&self, address: u16, count: u8) -> Option<Vec<u8>> {
+    pub fn read(&self, address: u16, count: u16) -> Option<Vec<u8>> {
         let mut buffer = vec![0; count as usize];
         if (self.read)(address, count, buffer.as_mut_ptr()) {
             Some(buffer)
@@ -19,7 +19,7 @@ impl StorageInterface {
     }
 
     pub fn write(&self, address: u16, bytes: &[u8]) -> Option<()> {
-        if (self.write)(address, bytes.len() as u8, bytes.as_ptr()) {
+        if (self.write)(address, bytes.len() as u16, bytes.as_ptr()) {
             Some(())
         } else {
             None
@@ -34,7 +34,7 @@ impl StorageInterface {
         let mut address = start;
         while bytes_remaining > 0 {
             let this_chunk_size = core::cmp::min(CHUNK_SIZE as u16, bytes_remaining);
-            if !(self.write)(address, this_chunk_size as u8, buffer.as_ptr()) {
+            if !(self.write)(address, this_chunk_size as u16, buffer.as_ptr()) {
                 return None;
             }
